Moving GetMatchingNodes API to search view model (#11146)
* Moving GetMatchingNodes API to search view model

* Clean Up

* Comments

* Create a dedicated test file to prepare for more unit tests
